ref(agent): Move node onboarding to its own file & clean-up
