+++ static/app/gettingStartedDocs/javascript/agentMonitoring.tsx
@@ -7,10 +7,11 @@ import type {
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {
-  getAgentMonitoringInstallStep,
-  getAgentMonitoringManualConfigStep,
-  getImport,
-} from 'sentry/gettingStartedDocs/node/utils';
+  getAgentIntegration,
+  getInstallStep,
+  getManualConfigureStep,
+} from 'sentry/gettingStartedDocs/node/agentMonitoring';
+import {getImport} from 'sentry/gettingStartedDocs/node/utils';
 import {t, tct} from 'sentry/locale';
 import {AgentIntegration} from 'sentry/views/insights/pages/agents/utils/agentIntegrations';
 
@@ -302,17 +303,16 @@ export function agentMonitoring({
 } = {}): OnboardingConfig {
   return {
     install: params =>
-      getAgentMonitoringInstallStep(params, {
+      getInstallStep(params, {
         packageName,
       }),
     configure: params => {
-      const selected =
-        (params.platformOptions as any)?.integration ?? AgentIntegration.VERCEL_AI;
+      const selected = getAgentIntegration(params);
 
       const importMode = 'esm-only';
 
       if (selected === AgentIntegration.MANUAL) {
-        return getAgentMonitoringManualConfigStep(params, {
+        return getManualConfigureStep(params, {
           packageName,
           importMode,
         });
