--- static/app/gettingStartedDocs/node-azurefunctions/agentMonitoring.tsx
@@ -1,3 +1,0 @@
-import {getNodeAgentMonitoringOnboarding} from 'sentry/gettingStartedDocs/node/utils';
-
-export const agentMonitoring = getNodeAgentMonitoringOnboarding();
