+++ static/app/gettingStartedDocs/node-azurefunctions/index.tsx
@@ -1,6 +1,6 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/node/agentMonitoring';
 
-import {agentMonitoring} from './agentMonitoring';
 import {crashReport} from './crashReport';
 import {logs} from './logs';
 import {mcp} from './mcp';
@@ -13,7 +13,7 @@ const docs: Docs = {
   crashReportOnboarding: crashReport,
   profilingOnboarding: profiling,
   logsOnboarding: logs,
-  agentMonitoringOnboarding: agentMonitoring,
+  agentMonitoringOnboarding: agentMonitoring(),
   mcpOnboarding: mcp,
   metricsOnboarding: metrics,
 };
