--- static/app/gettingStartedDocs/node-cloudflare-workers/agentMonitoring.tsx
@@ -1,5 +1,0 @@
-import {getNodeAgentMonitoringOnboarding} from 'sentry/gettingStartedDocs/node/utils';
-
-export const agentMonitoring = getNodeAgentMonitoringOnboarding({
-  packageName: '@sentry/cloudflare',
-});
