+++ static/app/gettingStartedDocs/node-cloudflare-workers/index.tsx
@@ -1,6 +1,6 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/node/agentMonitoring';
 
-import {agentMonitoring} from './agentMonitoring';
 import {crashReport} from './crashReport';
 import {logs} from './logs';
 import {mcp} from './mcp';
@@ -11,7 +11,9 @@ const docs: Docs = {
   onboarding,
   crashReportOnboarding: crashReport,
   logsOnboarding: logs,
-  agentMonitoringOnboarding: agentMonitoring,
+  agentMonitoringOnboarding: agentMonitoring({
+    packageName: '@sentry/cloudflare',
+  }),
   mcpOnboarding: mcp,
   metricsOnboarding: metrics,
 };
