+++ static/app/gettingStartedDocs/node-connect/index.tsx
@@ -1,6 +1,6 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/node/agentMonitoring';
 
-import {agentMonitoring} from './agentMonitoring';
 import {crashReport} from './crashReport';
 import {logs} from './logs';
 import {mcp} from './mcp';
@@ -14,7 +14,7 @@ const docs: Docs = {
   profilingOnboarding: profiling,
   logsOnboarding: logs,
   metricsOnboarding: metrics,
-  agentMonitoringOnboarding: agentMonitoring,
+  agentMonitoringOnboarding: agentMonitoring(),
   mcpOnboarding: mcp,
 };
