+++ static/app/gettingStartedDocs/node-express/index.tsx
@@ -3,8 +3,8 @@ import {
   feedbackOnboardingJsLoader,
   replayOnboardingJsLoader,
 } from 'sentry/gettingStartedDocs/javascript/jsLoader';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/node/agentMonitoring';
 
-import {agentMonitoring} from './agentMonitoring';
 import {crashReport} from './crashReport';
 import {logs} from './logs';
 import {mcp} from './mcp';
@@ -20,7 +20,7 @@ const docs: Docs = {
   logsOnboarding: logs,
   metricsOnboarding: metrics,
   profilingOnboarding: profiling,
-  agentMonitoringOnboarding: agentMonitoring,
+  agentMonitoringOnboarding: agentMonitoring(),
   mcpOnboarding: mcp,
 };
