--- static/app/gettingStartedDocs/node-gcpfunctions/agentMonitoring.tsx
@@ -1,5 +1,0 @@
-import {getNodeAgentMonitoringOnboarding} from 'sentry/gettingStartedDocs/node/utils';
-
-export const agentMonitoring = getNodeAgentMonitoringOnboarding({
-  packageName: '@sentry/google-cloud-serverless',
-});
